Move instrumentation specific classes out of javaagent-instrumentation-api
