+++ instrumentation/jaxrs/bootstrap/src/main/java/io/opentelemetry/javaagent/bootstrap/jaxrs/JaxrsContextPath.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.api.jaxrs;
+package io.opentelemetry.javaagent.bootstrap.jaxrs;
 
 import io.opentelemetry.context.Context;
 import io.opentelemetry.context.ContextKey;
@@ -12,7 +12,7 @@ import org.checkerframework.checker.nullness.qual.Nullable;
 /**
  * Helper container for storing context path for jax-rs requests. Jax-rs context path is the path
  * where jax-rs servlet is mapped or the value of ApplicationPath annotation. Span name is built by
- * combining servlet context path from {@link
+ * combining servlet context path from {@code
  * io.opentelemetry.instrumentation.api.servlet.ServletContextPath} jax-rs context path and the Path
  * annotation from called method or class.
  */
