+++ instrumentation/jaxrs/jaxrs-2.0/jaxrs-2.0-cxf-3.2/javaagent/build.gradle.kts
@@ -21,6 +21,8 @@ muzzle {
 }
 
 dependencies {
+  compileOnly(project(":instrumentation:jaxrs:bootstrap"))
+
   compileOnly("javax.ws.rs:javax.ws.rs-api:2.0")
   compileOnly("javax.servlet:javax.servlet-api:3.1.0")
   library("org.apache.cxf:cxf-rt-frontend-jaxrs:3.2.0")
