+++ instrumentation/jaxrs/jaxrs-2.0/jaxrs-2.0-resteasy-common/javaagent/build.gradle.kts
@@ -3,6 +3,8 @@ plugins {
 }
 
 dependencies {
+  compileOnly(project(":instrumentation:jaxrs:bootstrap"))
+
   compileOnly("javax.ws.rs:javax.ws.rs-api:2.0")
   compileOnly("org.jboss.resteasy:resteasy-jaxrs:3.1.0.Final")
