+++ settings.gradle.kts
@@ -157,6 +157,7 @@ include(":instrumentation:hibernate:hibernate-procedure-call-4.3:javaagent")
 include(":instrumentation:http-url-connection:javaagent")
 include(":instrumentation:hystrix-1.4:javaagent")
 include(":instrumentation:java-http-client:javaagent")
+include(":instrumentation:jaxrs:bootstrap")
 include(":instrumentation:jaxrs:jaxrs-1.0:javaagent")
 include(":instrumentation:jaxrs:jaxrs-2.0:jaxrs-2.0-arquillian-testing")
 include(":instrumentation:jaxrs:jaxrs-2.0:jaxrs-2.0-common:javaagent")
@@ -264,6 +265,7 @@ include(":instrumentation:reactor-netty:reactor-netty-0.9:javaagent")
 include(":instrumentation:reactor-netty:reactor-netty-1.0:javaagent")
 include(":instrumentation:rediscala-1.8:javaagent")
 include(":instrumentation:redisson-3.0:javaagent")
+include(":instrumentation:rmi:bootstrap")
 include(":instrumentation:rmi:javaagent")
 include(":instrumentation:rocketmq-client-4.8:javaagent")
 include(":instrumentation:rocketmq-client-4.8:library")
